big commit - upgraded to spring 1.2.5 - upgraded to hibernate 3.1b3 - added features for archetype service - added support for saveOrUpdate - added support for cascade all
svn path=/trunk/openvpms-framework/; revision=151

+++ openvpms-framework/src/java/org/openvpms/component/business/domain/im/lookup/LookupRelationship.java
@@ -44,7 +44,7 @@ public class LookupRelationship implements Serializable {
      * Uniquely identifies an instance
      */
     @SuppressWarnings("unused")
-    private int uid;
+    private long uid;
     
     /**
      * Used for database control
@@ -128,7 +128,7 @@ public class LookupRelationship implements Serializable {
     /**
      * @return Returns the uid.
      */
-    public int getUid() {
+    public long getUid() {
         return uid;
     }
